Re-added quest counter
