+++ Assets/Scripts/Quests/Quests.cs
@@ -11,10 +11,11 @@ namespace Quests
     public class Quests : MonoBehaviour
     {
         public static Action<Quest> OnQuestCompleted;
+        public static Action<Quest> OnQuestAdded;
+        public static Action<Quest> OnQuestRemoved;
         
         [SerializeField] private Transform dock, forestPath, mountainPath;
-        [SerializeField] private QuestCounter counter;
-        
+
         [SerializeField] private GameObject sectionPrefab;
         public GameObject SectionPrefab => sectionPrefab;
 
@@ -46,6 +47,7 @@ namespace Quests
             if (quests.Contains(q)) return false;
             quests.Add(q);
             q.Add();
+            OnQuestAdded?.Invoke(q);
             return true;
         }
 
@@ -54,6 +56,7 @@ namespace Quests
             if (!quests.Contains(q)) return false;
             quests.Remove(q);
             q.Remove();
+            OnQuestRemoved?.Invoke(q);
             return true;
         }
